mailman: add template for held messages on users list
Add a note to the held message template which explains that messages
held for being over the list size limit (default: 60K) will not be
approved.  Offer some suggestions to help senders reduce the size of
their messages.

This will save the list moderators the effort of doing this manually and
will prevent senders from waiting for a reply from those moderators.

References
==========
https://lists.fedoraproject.org/archives/list/users@lists.fedoraproject.org/thread/QIMCJNTLBEDJZEDLAQH53TBTP36VPG54/
https://lists.fedoraproject.org/archives/list/users@lists.fedoraproject.org/message/ZDVJHWAD64J3VKO5OQ2XITOOQMVL73GC/

Signed-off-by: Todd Zullinger <tmz@pobox.com>